+++ core/src/main/java/com/datastax/oss/sga/impl/agents/ai/GenAIToolKitFunctionAgentProvider.java
@@ -17,6 +17,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.function.Consumer;
 
 @Slf4j
 public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
@@ -25,7 +26,8 @@ public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
 
             "drop-fields", new StepConfigurationInitializer() {
                 @Override
-                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration) {
+                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration,
+                                          DataSourceConfigurationGenerator dataSourceConfigurationGenerator) {
                     requiredField(step, agentConfiguration, originalConfiguration, "fields");
                     optionalField(step, agentConfiguration, originalConfiguration, "part", null);
                 }
@@ -34,20 +36,23 @@ public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
             },
             "unwrap-key-value", new StepConfigurationInitializer() {
                 @Override
-                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration) {
+                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration,
+                                          DataSourceConfigurationGenerator dataSourceConfigurationGenerator) {
                     optionalField(step, agentConfiguration, originalConfiguration, "unwrapKey", null);
                 }
             },
             "cast", new StepConfigurationInitializer() {
                 @Override
-                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration) {
+                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration,
+                                          DataSourceConfigurationGenerator dataSourceConfigurationGenerator) {
                     requiredField(step, agentConfiguration, originalConfiguration, "schema-type");
                     optionalField(step, agentConfiguration, originalConfiguration, "part", null);
                 }
             },
             "flatten", new StepConfigurationInitializer() {
                 @Override
-                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration) {
+                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration,
+                                          DataSourceConfigurationGenerator dataSourceConfigurationGenerator) {
                     optionalField(step, agentConfiguration, originalConfiguration, "delimiter", null);
                     optionalField(step, agentConfiguration, originalConfiguration, "part", null);
                 }
@@ -56,13 +61,15 @@ public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
             },
             "compute", new StepConfigurationInitializer() {
                 @Override
-                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration) {
+                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration,
+                                          AgentConfiguration agentConfiguration, DataSourceConfigurationGenerator dataSourceConfigurationGenerator) {
                     requiredField(step, agentConfiguration, originalConfiguration, "fields");
                 }
             },
             "compute-ai-embeddings", new StepConfigurationInitializer() {
                 @Override
-                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration) {
+                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration,
+                                          DataSourceConfigurationGenerator dataSourceConfigurationGenerator) {
                     optionalField(step, agentConfiguration, originalConfiguration, "model", "text-embedding-ada-002");
                     requiredField(step, agentConfiguration, originalConfiguration, "embeddings-field");
                     requiredField(step, agentConfiguration, originalConfiguration, "text");
@@ -70,7 +77,18 @@ public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
             },
             "query", new StepConfigurationInitializer() {
                 @Override
-                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration) {
+                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration,
+                                          DataSourceConfigurationGenerator dataSourceConfigurationGenerator) {
+
+                    // reference to datasource
+                    String datasource = (String) originalConfiguration.remove("datasource");
+                    if (datasource == null) {
+                        // error
+                        requiredField(step, agentConfiguration, originalConfiguration, "datasource");
+                        return;
+                    }
+                    dataSourceConfigurationGenerator.generateDataSourceConfiguration(datasource);
+
                     requiredField(step, agentConfiguration, originalConfiguration, "fields");
                     requiredField(step, agentConfiguration, originalConfiguration, "query");
                     requiredField(step, agentConfiguration, originalConfiguration, "output-field");
@@ -78,7 +96,8 @@ public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
             },
             "chat-ai-completions", new StepConfigurationInitializer() {
                 @Override
-                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration) {
+                public void generateSteps(Map<String, Object> step, Map<String, Object> originalConfiguration, AgentConfiguration agentConfiguration,
+                                          DataSourceConfigurationGenerator dataSourceConfigurationGenerator) {
                     requiredField(step, agentConfiguration, originalConfiguration, "output-field");
                     requiredField(step, agentConfiguration, originalConfiguration, "messages");
                     requiredField(step, agentConfiguration, originalConfiguration, "model");
@@ -114,10 +133,16 @@ public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
     private interface StepConfigurationInitializer {
         default void generateSteps(Map<String, Object> step,
                            Map<String, Object> originalConfiguration,
-                           AgentConfiguration agentConfiguration) {}
+                           AgentConfiguration agentConfiguration,
+                           DataSourceConfigurationGenerator dataSourceConfigurationGenerator) {}
     }
 
-    protected void generateSteps(Map<String, Object> originalConfiguration, List<Map<String, Object>> steps, AgentConfiguration agentConfiguration) {
+    protected void generateSteps(Map<String, Object> originalConfiguration,
+                                 Map<String, Object> configuration,
+                                 Application application,
+                                 AgentConfiguration agentConfiguration) {
+        List<Map<String, Object>> steps = new ArrayList<>();
+        configuration.put("steps", steps);
         Map<String, Object> step = new HashMap<>();
 
         // we are mapping the original name to the ai-tools function name
@@ -126,45 +151,67 @@ public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
         // on every step you can put a "when" clause
         optionalField(step, agentConfiguration, originalConfiguration, "when", null);
 
+        DataSourceConfigurationGenerator dataSourceConfigurationInjector = (resourceId) -> {
+            generateDataSourceConfiguration(resourceId, application, configuration);
+        };
+
         STEP_TYPES.get(agentConfiguration.getType())
-                .generateSteps(step, originalConfiguration, agentConfiguration);
+                .generateSteps(step, originalConfiguration, agentConfiguration, dataSourceConfigurationInjector);
         steps.add(step);
     }
+
+    interface  DataSourceConfigurationGenerator {
+        void generateDataSourceConfiguration(String resourceId);
+    }
+
 
     private void generateOpenAIConfiguration(Application applicationInstance, Map<String, Object> configuration) {
         Resource resource = applicationInstance.getResources().values().stream()
                 .filter(r -> r.type().equals("open-ai-configuration"))
                 .findFirst().orElse(null);
         if (resource != null) {
-            String url = (String) resource.configuration().get("url");
-            String accessKey = (String) resource.configuration().get("access-key");
-            String provider = (String) resource.configuration().get("provider");
-            Map<String, Object> openaiConfiguration = new HashMap<>();
-            if (url != null) {
-                openaiConfiguration.put("url", url);
-            }
-            if (accessKey != null) {
-                openaiConfiguration.put("access-key", accessKey);
-            }
-            if (provider != null) {
-                openaiConfiguration.put("provider", provider);
-            }
+            Map<String, Object> openaiConfiguration = new HashMap<>(resource.configuration());
             configuration.put("openai", openaiConfiguration);
+        }
+    }
+
+    private void generateHuggingFaceConfiguration(Application applicationInstance, Map<String, Object> configuration) {
+        Resource resource = applicationInstance.getResources().values().stream()
+                .filter(r -> r.type().equals("hugging-face-configuration"))
+                .findFirst().orElse(null);
+        if (resource != null) {
+            Map<String, Object> huggingfaceConfiguration = new HashMap<>(resource.configuration());
+            configuration.put("huggingface", huggingfaceConfiguration);
+        }
+    }
+
+    private void generateDataSourceConfiguration(String resourceId, Application applicationInstance, Map<String, Object> configuration) {
+        Resource resource = applicationInstance.getResources().get(resourceId);
+        log.info("Generating datasource configuration for {}", resourceId);
+        if (resource != null) {
+            if (!resource.type().equals("datasource")) {
+                throw new IllegalArgumentException("Resource " + resourceId + " is not type=datasource");
+            }
+            if (configuration.containsKey("datasource")) {
+                throw new IllegalArgumentException("Only one datasource is supported");
+            }
+            configuration.put("datasource", new HashMap<>(resource.configuration()));
+        } else {
+            throw new IllegalArgumentException("Resource " + resourceId + " not found");
         }
     }
 
     @Override
     protected Map<String, Object> computeAgentConfiguration(AgentConfiguration agentConfiguration, Module module,
-                                                            ExecutionPlan physicalApplicationInstance,
+                                                            ExecutionPlan executionPlan,
                                                             ComputeClusterRuntime clusterRuntime) {
-        Map<String, Object> originalConfiguration = super.computeAgentConfiguration(agentConfiguration, module, physicalApplicationInstance, clusterRuntime);
+        Map<String, Object> originalConfiguration = super.computeAgentConfiguration(agentConfiguration, module, executionPlan, clusterRuntime);
         Map<String, Object> configuration = new HashMap<>();
 
-        generateOpenAIConfiguration(physicalApplicationInstance.getApplication(), configuration);
+        generateOpenAIConfiguration(executionPlan.getApplication(), configuration);
+        generateHuggingFaceConfiguration(executionPlan.getApplication(), configuration);
 
-        List<Map<String, Object>> steps = new ArrayList<>();
-        configuration.put("steps", steps);
-        generateSteps(originalConfiguration, steps, agentConfiguration);
+        generateSteps(originalConfiguration, configuration, executionPlan.getApplication(), agentConfiguration);
         return configuration;
     }
 
@@ -174,13 +221,28 @@ public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
         if (Objects.equals(previousAgent.getAgentType(), agentImplementation.getAgentType())
             && previousAgent instanceof DefaultAgentNode agent1
             && agentImplementation instanceof DefaultAgentNode agent2) {
-                Map<String, Object> configurationWithoutSteps1 = new HashMap<>(agent1.getConfiguration());
-                configurationWithoutSteps1.remove("steps");
-                Map<String, Object> configurationWithoutSteps2 = new HashMap<>(agent2.getConfiguration());
-                configurationWithoutSteps2.remove("steps");
-                log.info("Comparing {} and {}", configurationWithoutSteps1, configurationWithoutSteps2);
-                return configurationWithoutSteps1.equals(configurationWithoutSteps2);
-            }
+                Map<String, Object> purgedConfiguration1 = new HashMap<>(agent1.getConfiguration());
+                purgedConfiguration1.remove("steps");
+                Map<String, Object> purgedConfiguration2 = new HashMap<>(agent2.getConfiguration());
+                purgedConfiguration2.remove("steps");
+
+                // test query steps with different datasources
+                Object datasource1 = purgedConfiguration1.remove("datasource");
+                Object datasource2 = purgedConfiguration2.remove("datasource");
+
+                if (datasource1 != null && datasource2 != null) {
+                    log.info("Comparing datasources {} and {}", datasource1, datasource2);
+                     if (!Objects.equals(datasource1, datasource2)) {
+                         log.info("Agents {} and {} cannot be merged (different datasources)",
+                                 previousAgent, agentImplementation);
+                         return false;
+                     }
+                }
+
+                log.info("Comparing {} and {}", purgedConfiguration1, purgedConfiguration2);
+                return purgedConfiguration1.equals(purgedConfiguration2);
+        }
+        log.info("Agents {} and {} cannot be merged", previousAgent, agentImplementation);
         return false;
     }
 
@@ -201,8 +263,8 @@ public class GenAIToolKitFunctionAgentProvider extends AbstractAgentProvider {
             List<Map<String, Object>> steps2 = (List<Map<String, Object>>) configurationWithoutSteps2.remove("steps");
 
             List<Map<String, Object>> mergedSteps = new ArrayList<>();
-            mergedSteps.addAll(steps2);
             mergedSteps.addAll(steps1);
+            mergedSteps.addAll(steps2);
 
             Map<String, Object> result = new HashMap<>();
             result.putAll(configurationWithoutSteps1);
